Clean up demo
