+++ packages/demo/src/index.html
@@ -3,7 +3,9 @@
   <head>
     <meta charset="utf-8" />
     <title>My First Parcel App</title>
-    <script type="module" src="index.ts"></script>
+    <script type="module" src="./run.ts"></script>
+
+    </script>
   </head>
   <body>
     <div id="container"></div>
